Takes geth changes v1.11.14 -> v1.12.0 (#277)
* trie/

* remove rej file

* metrics/

* accounts/

* internal

* rpc/

* core/rawdb

* go.mod

* core/state

* eth/...

* core/vm core/types/ ethdb

* core/txpool

* left out

* additional core/

* should pass tests

* nits

* remove .rej file

* newTestBlockChain TODO

* Pr comments

* remove ExcessDataGas

* Add ExcessDataGas with verification

* Add verification in plugin/evm

+++ eth/gasprice/feehistory_test.go
@@ -44,9 +44,9 @@ import (
 func TestFeeHistory(t *testing.T) {
 	var cases = []struct {
 		pending      bool
-		maxCallBlock int
-		maxBlock     int
-		count        int
+		maxCallBlock uint64
+		maxBlock     uint64
+		count        uint64
 		last         rpc.BlockNumber
 		percent      []float64
 		expFirst     uint64
